Move SdkVersionPrefixHelper to configuration package
