+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/diagnostics/DiagnosticsHelper.java
@@ -5,6 +5,7 @@ package com.microsoft.applicationinsights.agent.internal.diagnostics;
 
 import com.azure.monitor.opentelemetry.exporter.implementation.statsbeat.RpAttachType;
 import com.azure.monitor.opentelemetry.exporter.implementation.utils.Strings;
+import com.microsoft.applicationinsights.agent.internal.configuration.SdkVersionPrefixHelper;
 import java.nio.file.Files;
 import java.nio.file.Path;
